Address reviewer feedback.
