+++ packages/core/src/core/coreToolScheduler.ts
@@ -594,6 +594,14 @@ export class CoreToolScheduler {
         const { request: reqInfo, invocation } = toolCall;
 
         try {
+          if (signal.aborted) {
+            this.setStatusInternal(
+              reqInfo.callId,
+              'cancelled',
+              'Tool call cancelled by user.',
+            );
+            continue;
+          }
           if (this.config.getApprovalMode() === ApprovalMode.YOLO) {
             this.setToolCallOutcome(
               reqInfo.callId,
@@ -601,14 +609,6 @@ export class CoreToolScheduler {
             );
             this.setStatusInternal(reqInfo.callId, 'scheduled');
           } else {
-            if (signal.aborted) {
-              this.setStatusInternal(
-                reqInfo.callId,
-                'cancelled',
-                'Tool call cancelled by user.',
-              );
-              continue;
-            }
             const confirmationDetails =
               await invocation.shouldConfirmExecute(signal);
